+++ core/api/test/integration/app/wallets/update-pending-invoices.spec.ts
@@ -1,7 +1,8 @@
 import { randomUUID } from "crypto"
 
 import { handleHeldInvoices } from "@app/wallets"
-import * as UpdatePendingInvoicesImpl from "@app/wallets/update-pending-invoices"
+import * as DeclinePendingInvoiceImpl from "@app/wallets/decline-single-pending-invoice"
+import * as UpdatePendingInvoiceImpl from "@app/wallets/update-single-pending-invoice"
 
 import { DEFAULT_EXPIRATIONS } from "@domain/bitcoin/lightning/invoice-expiration"
 import { WalletCurrency } from "@domain/shared"
@@ -20,7 +21,7 @@ describe("update pending invoices", () => {
       // Setup mocks
       const declineHeldInvoiceMock = jest.fn()
       const declineHeldInvoiceSpy = jest
-        .spyOn(UpdatePendingInvoicesImpl, "declineHeldInvoice")
+        .spyOn(DeclinePendingInvoiceImpl, "declineHeldInvoice")
         .mockImplementation(declineHeldInvoiceMock)
 
       // Setup expired USD wallet invoice
@@ -54,7 +55,9 @@ describe("update pending invoices", () => {
 
       // Expect declined invoice
       expect(declineHeldInvoiceMock.mock.calls.length).toBe(1)
-      expect(declineHeldInvoiceMock.mock.calls[0][0].paymentHash).toBe(paymentHash)
+      expect(declineHeldInvoiceMock.mock.calls[0][0].walletInvoice.paymentHash).toBe(
+        paymentHash,
+      )
 
       // Restore system state
       declineHeldInvoiceSpy.mockRestore()
@@ -64,12 +67,12 @@ describe("update pending invoices", () => {
       // Setup mocks
       const declineHeldInvoiceMock = jest.fn()
       const declineHeldInvoiceSpy = jest
-        .spyOn(UpdatePendingInvoicesImpl, "declineHeldInvoice")
+        .spyOn(DeclinePendingInvoiceImpl, "declineHeldInvoice")
         .mockImplementation(declineHeldInvoiceMock)
 
       const updatePendingInvoiceMock = jest.fn()
       const updatePendingInvoiceSpy = jest
-        .spyOn(UpdatePendingInvoicesImpl, "updatePendingInvoice")
+        .spyOn(UpdatePendingInvoiceImpl, "updatePendingInvoice")
         .mockImplementation(updatePendingInvoiceMock)
 
       // Setup expired BTC wallet invoice
